refactor: move QueueableEvent into `Mwop\App\EventDispatcher` namespace
