+++ src/App/EventDispatcher/QueueableEvent.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Mwop;
+namespace Mwop\App\EventDispatcher;
 
 use JsonSerializable;
